+++ kompendium-swagger-ui/src/test/kotlin/io/bkbn/kompendium/swagger/TestHelpers.kt
@@ -3,16 +3,22 @@ package io.bkbn.kompendium.swagger
 import io.bkbn.kompendium.core.Kompendium
 import io.bkbn.kompendium.core.fixtures.kompendium
 import io.kotest.assertions.ktor.shouldHaveStatus
+import io.kotest.matchers.Matcher
+import io.kotest.matchers.MatcherResult
+import io.kotest.matchers.should
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
 import io.kotest.matchers.string.shouldContain
 import io.ktor.application.Application
 import io.ktor.application.install
+import io.ktor.http.ContentType
 import io.ktor.http.HttpHeaders
 import io.ktor.http.HttpMethod
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.testing.TestApplicationEngine
+import io.ktor.server.testing.TestApplicationResponse
 import io.ktor.server.testing.createTestEnvironment
+import io.ktor.server.testing.contentType
 import io.ktor.server.testing.handleRequest
 import io.ktor.server.testing.withApplication
 import java.net.URI
@@ -39,7 +45,7 @@ object TestHelpers {
       realm: 'MY REALM',
       appName: 'TEST APP',
       useBasicAuthenticationWithAccessCodeGrant: true
-  });  
+  });
       """
             }
         )
@@ -52,10 +58,11 @@ object TestHelpers {
         }
     }
 
-    fun TestApplicationEngine.compareResource(resourceName: String, mustContain: List<String>) {
+    fun TestApplicationEngine.compareResource(resourceName: String, contentType: ContentType, mustContain: List<String>) {
         handleRequest(HttpMethod.Get, resourceName).apply {
             response shouldHaveStatus HttpStatusCode.OK
             response.content shouldNotBe null
+            response shouldHaveContentTypeMatching contentType
             mustContain.forEach {
                 response.content!! shouldContain it
             }
@@ -75,5 +82,17 @@ object TestHelpers {
             })
             test()
         }
+    }
+
+    private infix fun TestApplicationResponse.shouldHaveContentTypeMatching(contentType: ContentType) = this should haveContentTypeMatching(contentType)
+
+    private fun haveContentTypeMatching(contentType: ContentType) = object : Matcher<TestApplicationResponse> {
+      override fun test(value: TestApplicationResponse): MatcherResult {
+        return MatcherResult(
+          value.contentType().match(contentType),
+          { "Response should have ContentType $contentType= but was ${value.contentType()}" },
+          { "Response should not have ContentType $contentType" }
+        )
+      }
     }
 }